[Modify] Enable changes to search term
- One could now edit the current search term after changing the cursor
location using either the mouse or the arrow keys. If one wants to
navigate using the arrow keys after the picker is "listening" they can
now press the `Esc` key to continue navigating. In order to achieve this
susbstantial changes were made to how the search term is captured. This
required more visibility into what the "current word" being edited in
the input. There are some concerns with how much more intensive on the
browser this might be, so this should be tested under profiling when
the time comes.